requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pytesseract==0.3.10
opencv-python==4.8.1.78
pandas==2.1.4
numpy==1.26.2
spacy==3.7.2
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
regex==2023.10.3
